+++ src/test/java/woowacourse/documentation/AuthDocumentTest.java
@@ -1,0 +1,59 @@
+package woowacourse.documentation;
+
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyLong;
+import static org.mockito.Mockito.when;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessRequest;
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessResponse;
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.http.MediaType;
+import org.springframework.test.web.servlet.MockMvc;
+import woowacourse.auth.application.AuthService;
+import woowacourse.auth.dto.LoginRequest;
+import woowacourse.auth.dto.TokenResponse;
+import woowacourse.auth.support.JwtTokenProvider;
+
+@AutoConfigureMockMvc
+@AutoConfigureRestDocs
+@SpringBootTest
+public class AuthDocumentTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+    @Autowired
+    private ObjectMapper objectMapper;
+    @Autowired
+    private JwtTokenProvider jwtTokenProvider;
+    @MockBean
+    private AuthService authService;
+
+    @Test
+    void login() throws Exception {
+        when(authService.loginCustomer(any(LoginRequest.class))).thenReturn(1L);
+
+        TokenResponse tokenResponse = new TokenResponse(jwtTokenProvider.createToken(String.valueOf(1L)));
+        when(authService.createToken(anyLong())).thenReturn(tokenResponse);
+
+        LoginRequest loginRequest = new LoginRequest("email@woowahan.com", "Pw123456!");
+
+        this.mockMvc.perform(post("/auth/login")
+                        .content(objectMapper.writeValueAsString(loginRequest))
+                        .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().isOk())
+                .andDo(document("auth/login",
+                        preprocessRequest(prettyPrint()),
+                        preprocessResponse(prettyPrint())
+                ));
+    }
+}
