+++ src/test/java/woowacourse/documentation/OrderDocumentTest.java
@@ -1,17 +1,19 @@
-package woowacourse.document;
+package woowacourse.documentation;
 
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.Mockito.when;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
-import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessRequest;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessResponse;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import org.mockito.ArgumentMatchers;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -20,16 +22,15 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import woowacourse.auth.support.JwtTokenProvider;
-import woowacourse.auth.support.User;
-import woowacourse.shoppingcart.application.ProductService;
+import woowacourse.shoppingcart.application.OrderService;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.dto.ProductResponse;
-import woowacourse.shoppingcart.dto.ProductsResponse;
+import woowacourse.shoppingcart.dto.OrderRequest;
+import woowacourse.shoppingcart.dto.OrderResponse;
 
 @AutoConfigureMockMvc
 @AutoConfigureRestDocs
 @SpringBootTest
-class ProductDocumentTest {
+class OrderDocumentTest {
 
     private final Product product1 = new Product(
             1L, "콜드 브루 몰트",
@@ -42,41 +43,61 @@ class ProductDocumentTest {
     @Autowired
     private MockMvc mockMvc;
     @Autowired
+    private ObjectMapper objectMapper;
+    @Autowired
     private JwtTokenProvider jwtTokenProvider;
     @MockBean
-    private ProductService productService;
+    private OrderService orderService;
 
     @Test
-    void findWithMember() throws Exception {
-        List<ProductResponse> productResponses = List.of(
-                ProductResponse.from(product1, 5),
-                ProductResponse.from(product2, 10)
-        );
-        when(productService.findProducts(any(User.class))).thenReturn(new ProductsResponse(productResponses));
+    void orderItem() throws Exception {
+        when(orderService.addOrder(anyLong(), ArgumentMatchers.any(OrderRequest.class))).thenReturn(1L);
 
-        String token = jwtTokenProvider.createToken(String.valueOf(anyLong()));
-        this.mockMvc.perform(get("/products")
+        OrderRequest orderRequest = new OrderRequest(1L, 2);
+        String token = jwtTokenProvider.createToken(String.valueOf(1L));
+
+        this.mockMvc.perform(post("/customers/orders")
                         .header("Authorization", "Bearer " + token)
+                        .content(objectMapper.writeValueAsString(orderRequest))
+                        .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk())
-                .andDo(document("product/find/member",
+                .andExpect(status().isCreated())
+                .andDo(document("order/create",
                         preprocessRequest(prettyPrint()),
                         preprocessResponse(prettyPrint())
                 ));
     }
 
     @Test
-    void findWithNonMember() throws Exception {
-        List<ProductResponse> productResponses = List.of(
-                ProductResponse.from(product1, 0),
-                ProductResponse.from(product2, 0)
+    void findAll() throws Exception {
+        List<OrderResponse> orderResponses = List.of(
+                OrderResponse.from(1L, product1, 3),
+                OrderResponse.from(2L, product2, 15)
         );
-        when(productService.findProducts(any(User.class))).thenReturn(new ProductsResponse(productResponses));
+        when(orderService.findOrdersByCustomerId(anyLong())).thenReturn(orderResponses);
 
-        this.mockMvc.perform(get("/products")
+        String token = jwtTokenProvider.createToken(String.valueOf(anyLong()));
+        this.mockMvc.perform(get("/customers/orders")
+                        .header("Authorization", "Bearer " + token)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
-                .andDo(document("product/find/anonymous",
+                .andDo(document("order/find/all",
+                        preprocessRequest(prettyPrint()),
+                        preprocessResponse(prettyPrint())
+                ));
+    }
+
+    @Test
+    void findOrder() throws Exception {
+        OrderResponse orderResponse = OrderResponse.from(1L, product1, 20);
+        when(orderService.findOrderById(anyLong(), anyLong())).thenReturn(orderResponse);
+
+        String token = jwtTokenProvider.createToken(String.valueOf(1L));
+        this.mockMvc.perform(get("/customers/orders/{orderId}", 1L)
+                        .header("Authorization", "Bearer " + token)
+                        .accept(MediaType.APPLICATION_JSON))
+                .andExpect(status().isOk())
+                .andDo(document("order/find",
                         preprocessRequest(prettyPrint()),
                         preprocessResponse(prettyPrint())
                 ));
